\name{test_regression_stump_R}
\alias{test_regression_stump_R}
\docType{data}
\title{
Test the regression model
}
\description{
Test the regression model for each target gene
}
%\usage{data("test_regression_stump_R")}
\format{
  The format is:
List of 4
 $ name         : chr "test_regression_stump_R"
 $ address      :Class 'RegisteredNativeSymbol' <externalptr> 
 $ dll          :List of 5
  ..$ name         : chr "RGBM"
  ..$ path         : chr "/home/raghvendra/R/x86_64-pc-linux-gnu-library/3.3/RGBM/libs/RGBM.so"
  ..$ dynamicLookup: logi TRUE
  ..$ handle       :Class 'DLLHandle' <externalptr> 
  ..$ info         :Class 'DLLInfoReference' <externalptr> 
  ..- attr(*, "class")= chr "DLLInfo"
 $ numParameters: int 15
 - attr(*, "class")= chr [1:2] "CRoutine" "NativeSymbolInfo"
}
\author{
Raghvendra Mall <rmall@hbku.edu.qa>
}